+++ src/Mibo.Core.Tests/Main.fs
@@ -1,0 +1,7 @@
+module Mibo.Core.Tests.Main
+
+open Expecto
+
+[<EntryPoint>]
+let main argv =
+  Tests.runTestsInAssemblyWithCLIArgs [] argv