+++ src/Mibo.Core/Properties/AssemblyInfo.fs
@@ -1,0 +1,9 @@
+namespace Mibo.Core.AssemblyInfo
+
+open System.Runtime.CompilerServices
+
+[<assembly: InternalsVisibleTo("Mibo.Core.Tests")>]
+[<assembly: InternalsVisibleTo("Mibo.Raylib")>]
+[<assembly: InternalsVisibleTo("Mibo.Raylib.Tests")>]
+[<assembly: InternalsVisibleTo("Mibo.MonoGame")>]
+do ()